rename routes folder to resources
